@@ -30,7 +30,6 @@
     "grid",
     "gridLayout",
     "angular",
-    "angular2",
     "dashboard",
     "ngx",
     "drag&drop",
@@ -41,38 +40,37 @@
   },
   "homepage": "https://github.com/swiety85/angular2gridster#readme",
   "dependencies": {
-    "@angular/common": "^7.0.2",
-    "@angular/core": "^7.0.2",
-    "@angular/compiler": "^7.0.2",
-    "@angular/animations": "^7.0.2",
-    "@angular/forms": "^7.0.2",
-    "@angular/http": "^7.0.2",
-    "@angular/platform-browser": "^7.0.2",
-    "@angular/platform-browser-dynamic": "^7.0.2",
-    "@angular/router": "^7.0.2",
+    "@angular/common": "^8.0.0",
+    "@angular/core": "^8.0.0",
+    "@angular/compiler": "^8.0.0",
+    "@angular/animations": "^8.0.0",
+    "@angular/forms": "^8.0.0",
+    "@angular/platform-browser": "^8.0.0",
+    "@angular/platform-browser-dynamic": "^8.0.0",
+    "@angular/router": "^8.0.0",
     "core-js": "^2.5.4",
     "rxjs": "^6.3.3",
-    "zone.js": "^0.8.26"
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.10.3",
-    "@angular-devkit/build-ng-packagr": "~0.10.3",
-    "@angular/cli": "~7.0.3",
-    "@angular/compiler-cli": "^7.0.2",
-    "@angular/language-service": "^7.0.2",
+    "@angular-devkit/build-angular": "~0.800.0",
+    "@angular-devkit/build-ng-packagr": "~0.800.0",
+    "@angular/cli": "~8.0.2",
+    "@angular/compiler-cli": "^8.0.0",
+    "@angular/language-service": "^8.0.0",
     "@types/jasmine": "~2.8.9",
     "@types/jasminewd2": "~2.0.5",
     "@types/node": "~10.12.2",
-    "codelyzer": "~4.5.0",
+    "codelyzer": "^5.0.1",
     "edit-json-file": "^1.1.0",
     "gh-pages": "^2.0.1",
-    "ng-packagr": "^4.4.0",
-    "standard-version": "^4.4.0",
+    "ng-packagr": "^5.1.0",
+    "standard-version": "^6.0.1",
     "ts-node": "~7.0.1",
-    "tsickle": "0.33.1",
-    "tslib": "^1.9.3",
+    "tsickle": "^0.35.0",
+    "tslib": "^1.9.0",
     "tslint": "~5.11.0",
-    "typescript": "~3.1.5"
+    "typescript": "~3.4.5"
   },
   "peerDependencies": {}
 }
