@@ -1,7 +1,7 @@
 {
     "name": "angular2gridster",
     "peerDependencies": {
-        "@angular/common": "^7.0.0",
-        "@angular/core": "^7.0.0"
+        "@angular/common": "^8.0.0",
+        "@angular/core": "^8.0.0"
     }
 }
